page_store: split BufferSet into separate file
